requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
marshmallow==3.19.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pdfminer.six==20221105
tabula-py==2.7.0
pandas==1.5.3
bcrypt==4.0.1
Werkzeug==2.2.3
email-validator==2.0.0
cryptography==39.0.2
gunicorn==20.1.0 